Add NewTestServers helper function to create multiple servers in e2e
